Update player making move logic, made game playable
Added and renamed methods so that 2 players are able to play
chess.

Methods added and changed:
1. #play
2. #player_make_move
3. #execute_move
4. #choose_move
5. #display_available_moves
6. #piece_can_be_played
7. #array_position_to_coordinates

Need to add a go back choice in the #choose_move
method so that the player is able to see what moves
are available to a specific piece and then go back
to play a different piece.