Add #filter_moves_for_check, still needs work
If piece is a King it removes moves that would make
the King in check.

The #collect_enemy_moves needs to be updated and renamed
to #collect_enemy_attack_moves so that it collects only
attack moves. That means that pawn regular moves need to
be ignored and attack moves should be taken instead.

The #filter_moves_for_check should also check if the specific
piece is moved, will it make the king be in check and
should accordingly remove such moves.