Fix bug in #get_moves Pawn class
Before if you tried to move a pawn from a starting
position by 2 spaces and there was a piece blocking
the way. It was simply ignored and the pawn was
allowed to move over it.

Now this is fixed, the pawn is no longer allowed to
jump over pieces from the starting position.