Update castling, move works but needs restrictions
Before the castling move only partially worked.
The #execute_move method would change a
piece's moved state from false to true. This method
was being used to simulate every enemy move available
to them in #move_puts_king_in_check? and #filter_moves_for_check
methods.

This would make the enemy not eligible for castling
even if they met the other conditions and did not
move the king or the rook.

Also made it so that the rook would change positions
after the king made the castling move.

Still needs restrictions for if the enemy is attacking
the squares through which the king would be castling
through.