Add a simple AI, fix en_passant bug in #execute_move
Added a simple AI that plays as the black team if
the player chooses to do that. The AI picks a random
piece that can be moved, from that piece picks a
random move and then executes it.

Found a bug where a knight move would count as an
en_passant.

Fixed by adding a is_a?(Pawn) check in the #is_an_en_passant_move?
method.