Delete nonfunctional chess pieces test
Because the board that is passed to the chess pieces
is no longer a 2D array containing string representations
of pieces, but contains the actual pieces the test
written before no longer work.

They will be partially rewritten in the chess class
testing. Instead of testing the #get_moves in pieces,
test will be written for the #generate_legal_moves,
which will take pieces, their positions and color as
arguments and return all legal moves available to that
piece which will effectively replace the previous test.